extract segments logic as a separate file
